refactor(graphcache): split cache lifecycle into focused files
WIP snapshot: extract gc, traversal, write, indexes, stats, causality,
and doc out of cache.go; include issue-732 bench reports.
