+++ core/graphcache/head_index.go
@@ -86,73 +86,3 @@ func (h *headIndex) walkPrefix(prefix string, fn func(projected string, headID v
 		return true
 	})
 }
-
-// --- GraphCache wiring -----------------------------------------------------
-//
-// The following helpers are mounted on GraphCache and called only by the
-// edge write paths (single + batch + flush). They preserve a single
-// invariant: when c.headByTail != nil, every edge present in c.edges has
-// a matching (projected, headID) entry in c.headByTail[tailID], and vice
-// versa. Callers MUST hold c.mu in write mode (Lock, not RLock).
-
-// onEdgeAddedLocked maintains the per-tail head index after a successful
-// edges.addWithExpiration. When created is false the underlying bucket
-// already existed and the index already reflects it; the call is a no-op.
-func (c *GraphCache[S, T]) onEdgeAddedLocked(created bool, tailID, headID vertexID, head S) {
-	if !created || c.headByTail == nil {
-		return
-	}
-	hi := c.headByTail[tailID]
-	if hi == nil {
-		hi = newHeadIndex()
-		c.headByTail[tailID] = hi
-	}
-	hi.insert(headID, c.prefixExtract(head))
-}
-
-// onEdgeDeletedLocked maintains the per-tail head index after a successful
-// edges.delete (or before a flushFunc deleteLocked). The head argument is
-// the original S key the caller already has in hand, so we never have to
-// resolve it through the dict — which is important because dict refs are
-// released by deleteLocked.
-func (c *GraphCache[S, T]) onEdgeDeletedLocked(tailID, headID vertexID, head S) {
-	if c.headByTail == nil {
-		return
-	}
-	hi, ok := c.headByTail[tailID]
-	if !ok {
-		return
-	}
-	if empty := hi.delete(headID, c.prefixExtract(head)); empty {
-		delete(c.headByTail, tailID)
-	}
-}
-
-// headIndexOnFlushDeleteLocked returns the onDelete callback handed to
-// edgeCache.flushFunc. It captures head S BEFORE deleteLocked runs
-// (flushFunc invokes the callback first), so resolveID is guaranteed to
-// succeed. Returns nil when the head index is disabled, in which case
-// flushFunc skips the call entirely.
-func (c *GraphCache[S, T]) headIndexOnFlushDeleteLocked() func(tailID, headID vertexID) {
-	if c.headByTail == nil {
-		return nil
-	}
-	return func(tailID, headID vertexID) {
-		head, ok := c.edges.resolveID(headID)
-		if !ok {
-			return
-		}
-		c.onEdgeDeletedLocked(tailID, headID, head)
-	}
-}
-
-// disableHeadIndexForTesting drops the per-tail head index so
-// ScanEdgesByPrefix falls back to the v1 materialise-and-sort path. It
-// exists so regression tests and before/after benchmarks can exercise
-// the fallback under identical data. Not part of the public API: lower
-// case on purpose, and never invoked by production code.
-func (c *GraphCache[S, T]) disableHeadIndexForTesting() {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.headByTail = nil
-}
