+++ testbed/bench/scenarios/replication_apply_churn.yaml
@@ -19,7 +19,6 @@
 # exactly the surface #700 leaked on. snapshot_runtime samples all three.
 name: replication_apply_churn
 cluster:
-  default_ttl_seconds: 15
   gc_interval_seconds: 10
   # Disable the search index: this scenario isolates the vertexHLC apply-path
   # heap signal. Leaving search on would add the InvertedIndex's own churn
@@ -34,19 +33,19 @@ target:
   endpoints: ["localhost:6380"]
   calls:
     # [0] PutVertex via the replicated-write path. A bounded, churned key
-    #     space (mod 10000, rewritten faster than the 15s TTL) keeps the
-    #     apply path continuously setting + sweeping vertexHLC at a bounded
-    #     live cardinality. Bounded value vocabulary (mod 100) keeps any
-    #     incidental indexing bounded.
+    #     space (mod 30000, rewritten slower than the 15s TTL at 1.5k
+    #     writes/s) keeps the apply path continuously setting + sweeping
+    #     vertexHLC at a bounded live cardinality. Bounded value vocabulary
+    #     (mod 100) keeps any incidental indexing bounded.
     - call: graph.v1.LanternService/PutVertex
       data_template: |
-        {"vertex":{"key":"hlc-churn-{{mod .RequestNumber 10000}}","string":"v{{mod .RequestNumber 100}}"}}
+        {"vertex":{"key":"hlc-churn-{{mod .RequestNumber 30000}}","string":"v{{mod .RequestNumber 100}}","expiration":"{{ now | dateModify `15s` | date `2006-01-02T15:04:05.999999999Z07:00` }}"}}
     # [1] Plural read over the same key space — returns OK + `missing` (never
     #     NotFound) so reads add load without inflating the non-OK count as
     #     keys age out between rewrites.
     - call: graph.v1.LanternService/GetVertices
       data_template: |
-        {"keys":["hlc-churn-{{mod .RequestNumber 10000}}"]}
+        {"keys":["hlc-churn-{{mod .RequestNumber 30000}}"]}
 leak_gate:
   goroutine_max_delta: 30
   heap_alloc_max_delta_mb: 64
